Refactor all four reminder lists implementations
