+++ src/client/modules/Reminders/InvestmentsNoRecentInteractionsList.jsx
@@ -21,11 +21,12 @@ import {
 
 import { sortOptions, maxItemsToPaginate, itemsPerPage } from './constants'
 
+import InvestmentItemRenderer from './ItemRenderers/InvestmentItemRenderer'
+import { CollectionSort, RoutedPagination } from '../../components'
 import CollectionHeader from './CollectionHeader'
+import CollectionList from './CollectionList'
 import Effect from '../../components/Effect'
 import Task from '../../components/Task'
-import InvestmentsCollectionList from './InvestmentsCollectionList'
-import { CollectionSort, RoutedPagination } from '../../components'
 
 const Summary = styled('p')({
   color: BLACK,
@@ -92,8 +93,9 @@ const InvestmentsNoRecentInteractionsList = ({
                       })
                     }
                   />
-                  <InvestmentsCollectionList
+                  <CollectionList
                     results={results}
+                    itemRenderer={InvestmentItemRenderer}
                     disableDelete={deleteTask.status || nextPending}
                     onDeleteReminder={(reminderId) => {
                       deleteTask.start({
